+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Added
 - Handle thrown JS runtime custom exceptions containing a message and a grpc code to be returned in the server response.
 - Add runtime function to retrieve a random set of users.
+- Add runtime ChannelMessageSend functions.
 
 ### Changed
 - Size limit for status messages increased from 128 to 2048 characters.
@@ -24,6 +25,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Fix JS runtime missing fields from leaderboards/tournaments get, list and write functions.
 - Fix JS runtime ownerId field not working correctly in leaderboard/tournament records list functions.
 - Fix parameter usage in leaderboard score set operator.
+- Fix JS runtime storageList not returning the cursor.
 
 ## [3.4.0] - 2021-07-08
 ### Added